Creates new compoent Home and moves redux to App
